refactor: move run function to lib
